[issue 259] make fields final, break out StandardLauncher as per github review
